kmeans clustering with gap statistics
To run, just use (example) command: python "locationOfFile.../kmeans.py"
"locationOfQueryDataFolder.../QueryData/1000iterations/NP2004.data" 5 10

if you want to set k in the range 5 to 10.

The data gets saved to disc, in a folder that gets created (if it isn't
already there) named ClusterData which is three directories up from
where the script is run. Obviously the class clusterData.py needs to be
in the same directory as kmeans2.py in order to import the class and
create the objects when running

remember, the objects return types are:
- dict[clusterNumber:int] = list of all rankers (where rankers are also
lists)
- dict[queryID:string] = list of cluster numbers as ints

if you want to run it in eclipse or sth there is also in the code a line
(line 213) that i've commented out, just above the one that it used to
take arguments from command prompt.